EX-Daily expense app with user Authentication

+++ controllers/expense.js
@@ -1,10 +1,9 @@
 const path = require("path");
 
-const rootdir = require("../util/path.js");
 const expense = require("../models/expense.js");
 
 exports.expense = (req, res, next) => {
-  res.sendFile(path.join(rootdir, "views", "user", "expense", "expense.html"));
+  res.sendFile(path.join(__dirname, "..", "views", "expense.html"));
 };
 
 exports.addExpense = async (req, res, next) => {
@@ -15,6 +14,7 @@ exports.addExpense = async (req, res, next) => {
       Amount: amount,
       Description: description,
       Category: category,
+      userId: req.user.id
     });
 
     res.status(201).json({ expenseDetails: data });
@@ -27,7 +27,7 @@ exports.addExpense = async (req, res, next) => {
 
 exports.getExpense = async (req, res, next) => {
   try {
-    const details = await expense.findAll();
+    const details = await expense.findAll({ where: { userId: req.user.id } });
     res.status(201).json({ allExpenseDetails: details });
   } catch (err) {
     res.status(500).json({
@@ -39,7 +39,7 @@ exports.getExpense = async (req, res, next) => {
 exports.deleteExpense = async (req, res, next) => {
   try {
     const id = req.params.id;
-    await expense.destroy({ where: { id: id } });
+    await expense.destroy({ where: { id: id, userId: req.user.id } });
     res.sendStatus(200);
   } catch (err) {
     res.status(500).json({
